func_blocks_analysis/requirements.txt
# Core Data Processing
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0,<1.12.0

# Machine Learning
scikit-learn>=1.3.0,<1.4.0

# Data Profiling
ydata-profiling>=4.5.0
sweetviz>=2.3.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<0.14.0
plotly>=6.1.1,<7.0.0
kaleido==0.2.1

# Sequential Pattern Mining
mlxtend>=0.23.0,<0.24.0

# Graph Analysis and Network Visualization
networkx>=3.0.0,<4.0.0
pyvis>=0.3.2

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0,<9.0.0
notebook>=7.0.0

# Data Storage
pyarrow>=14.0.0,<15.0.0
fastparquet>=2023.10.0

# Configuration
pyyaml>=6.0.0

# Progress Bars
tqdm>=4.65.0

# Utils
python-dateutil>=2.8.0

# Excel export (optional)
openpyxl>=3.1.0

# PDF export (optional - выберите один)
# playwright>=1.40.0  # Рекомендуется для PDF
# pdfkit>=1.0.0       # Альтернатива (требует wkhtmltopdf)
# weasyprint>=60.0    # Альтернатива

# Additional utilities
humanize>=4.9.0  # Для красивого форматирования чисел
colorlog>=6.8.0  # Для цветного логирования